+++ samlang-core-compiler/index.ts
@@ -1,5 +1,5 @@
 import compileSamlangSourcesToHighIRSources from './hir-toplevel-lowering';
-import lowerMidIRSourcesToLLVMSources from './llvm-lowering-translator';
+import lowerMidIRSourcesToLLVMSources from './llvm-sources-lowering';
 import lowerHighIRSourcesToMidIRSources from './mir-sources-lowering';
 
 export {
